audio-kernel: aw882xx: Remove build timestamp injection
This causes parts of the audio module to be rebuilt during every
incremental build, even if there are no changes:

  CC      techpack/audio/ipc/apr.o - due to command line change
  CC      techpack/audio/ipc/apr_v2.o - due to command line change
  CC      techpack/audio/ipc/apr_tal_rpmsg.o - due to command line change
  CC      techpack/audio/ipc/wcd-dsp-glink.o - due to command line change

We're only experiencing this issue in techpack/audio/ipc at the moment,
but kill the timestamp injection in all the audio components to
eliminate the possibility of encountering this issue again in the
future. This is harmless since the injected BUILD_TIMESTAMP macro is
never used.

Change-Id: I35b588a2a902438b2c79ebbac31453f2e528bbbe
Signed-off-by: Danny Lin <danny@kdrag0n.dev>